huachenghui/requirement.txt
certifi==2019.6.16
cffi==1.10.0
chardet==3.0.4
cycler==0.10.0
Django==2.1.10
django-cors-headers==3.0.2
djangorestframework==3.9.4
idna==2.8
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.16.4
olefile==0.44
opencv-python==4.1.0.25
Pillow==6.1.0
pycparser==2.18
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
scipy==1.3.0
six==1.12.0
sqlparse==0.3.0
torch==1.1.0
torchvision==0.3.0
tqdm==4.32.2
urllib3==1.25.3
